Python estudos

+++ Exercícios Práticos/ex4.py
@@ -1,5 +1,5 @@
 def main():
-    numero = float(input("Digite um número décimal: "))
+    numero = int(input("Digite um número décimal: "))
 
     print(bin(numero))
     print(oct(numero))
